# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "IsoString": {
        "type": "string"
      },
...
      "HistoricalFundingResponseObject": {
        "properties": {
          "ticker": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "effectiveAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "effectiveAtHeight": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "rate",
          "price",
          "effectiveAt",
          "effectiveAtHeight"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "HistoricalFundingResponse": {
        "properties": {
          "historicalFunding": {
            "items": {
              "$ref": "#/components/schemas/HistoricalFundingResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "historicalFunding"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "SubaccountHistoricalFundingResponseObject": {
        "properties": {
          "market": {
            "type": "string"
          },
          "positionType": {
            "type": "string",
            "enum": [
              "LONG",
              "SHORT"
            ]
          },
          "date": {
            "$ref": "#/components/schemas/IsoString"
          },
          "positionSize": {
            "type": "string"
          },
          "payment": {
            "type": "string"
          },
          "fundingRate": {
            "type": "string"
          }
        },
        "required": [
          "market",
          "positionType",
          "date",
          "positionSize",
          "payment",
          "fundingRate"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "SubaccountHistoricalFundingResponse": {
        "properties": {
          "historicalFunding": {
            "items": {
              "$ref": "#/components/schemas/SubaccountHistoricalFundingResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "historicalFunding"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Indexer API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/historicalFunding/{ticker}": {
      "get": {
        "operationId": "GetHistoricalFunding",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalFundingResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
    "/historicalFunding/address/{address}/{subaccountNumber}": {
      "get": {
        "operationId": "GetSubaccountHistoricalFunding",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubaccountHistoricalFundingResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "effectiveBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "",
      "description": "Public Testnet"
    }
  ]
}